LayoutTests/js/dom/promise-rejection-might-stack-overflow-expected.txt

CONSOLE MESSAGE: RangeError: Maximum call stack size exceeded.
PASS Got expected error: 'RangeError: Maximum call stack size exceeded.'
PASS successfullyParsed is true

TEST COMPLETE
